requirements.txt
aiofiles==0.8.0
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
colored==1.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
dict-toolbox==2.3.0
distlib==0.3.5
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
filelock==3.7.1
identify==2.5.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.8.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
msgpack==1.0.4
nest-asyncio==1.5.5
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pop==21.0.2
pop-config==9.0.0
pop-loop==1.0.4
pre-commit==2.20.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redgreenunittest==0.1.1
rend==6.4.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.2
